Made the conatcs list work (Read description)
I made a new function on line 318 it is a direct copy of the searchContacts function the only difference is that it calls the initContactsTable function directly after the contactsList array is created. This works around all the scheduling issues we were having but loses the reusability of the search function.